Adding gdrive spreadsheet functionality, keeping lowest price only
With every run of price tracker, the script will write data to a Google Spreasheeat. Spreadsheet "Amazon Price Tracker" must be created manually by the user.

This commit also changes from storing data in redis as old_price:new_price to a model of lowest_price:new_price. Lowest price will be lowest since the tracker was run for the first time. I'll need to test it more to see if this or previous model make more sense.

Fixes #1